- ommit pnm lines \n- change messages\n
git-svn-id: http://pkgbuild.svn.sourceforge.net/svnroot/pkgbuild/spec-files-extra/trunk@3389 3354ce84-6216-0410-ad3d-afc617165c18

+++ experimental/get_pnm_names.sh
@@ -21,14 +21,18 @@ function get_ips_name_list {
     typeset SVRPKGS=$*
 
     typeset TEMP=`mktemp /tmp/get_ips_name_list.XXXXXX`
+    rm -f ${TEMP}
     touch ${TEMP}
 
-    
     for pkgname in $SVRPKGS;do
-	echo searching ${pkgname}' ... \c' > /dev/stderr
-	IPSNAME=`get_ips_name ${pkgname}`
-	echo ${IPSNAME}>/dev/stderr
-	echo ${pkgname} ${IPSNAME}>> ${TEMP}
+	if echo $pkgname | egrep -e 'pnm_(build)?requires_' > /dev/null 2>&1; then
+	    echo "$pkgname is already changes">/dev/stderr
+	else
+	    echo searching ${pkgname}' ... \c' > /dev/stderr
+	    IPSNAME=`get_ips_name ${pkgname}`
+	    echo ${IPSNAME} >/dev/stderr
+	    echo ${pkgname} ${IPSNAME}>> ${TEMP}
+	fi
     done
     cat ${TEMP}
     rm -f ${TEMP}
@@ -54,11 +58,12 @@ BREQ_TEMP=`mktemp /tmp/make_pnm_macros_build_require.XXXXXX`
 get_ips_name_list ${REQUIRES} >${REQ_TEMP}
 get_ips_name_list ${BUILD_REQUIRES} > ${BREQ_TEMP}
 
-echo --- Cut Here, to include/packagenames.define*.inc by your hand, if you need ----
+echo ----- These package names are exist in include/packagenames.define*.inc, perfectly ?---
+echo ----- You need to check include/packagenames.define*.inc. -----------------------------
 cat ${REQ_TEMP} | sed 's/^\([^ ]*\)[\ \t]*pkg:\/\([^@]*\)@.*$/%define pnm_requires_\1 \2/' | sed 's/-/_/g'
 cat ${BREQ_TEMP}| sed 's/^\([^ ]*\)[\ \t]*pkg:\/\([^@]*\)@.*$/%define pnm_buildrequires_\1 \2/' | sed 's/-/_/g'
 
-echo ------ Cut Here, to ${SPECFILE} by your hand -----------------------------------
+echo ------ Cut Here, and Change to ${SPECFILE} by your hand -----------------------------------
 cat ${REQ_TEMP} | awk '{printf"Requires: %%{pnm_requires_%s}\n",$1}' | sed 's/-/_/g'
 cat ${BREQ_TEMP}| awk '{printf"BuildRequires: %%{pnm_buildrequires_%s}\n",$1}' | sed 's/-/_/g'
